WIP checkin

+++ api_inventory.py
@@ -7,17 +7,20 @@ from constants import *
 # VenntHandler methods
 
 def view_items(self, args, username):
-	# TODO assert username has permissions
+	if self.server.db.permissions(username, character_id) < Permission.PRIVATE_VIEW:
+		return self.respond({"success":False, "info":MSG_NO_PERMISSION})
 	items = self.server.db.view_items(username, args[KEY_ID])
 	return self.respond({"success":True, "value":items})
 	
 def remove_item(self, args, username):
-	# TODO assert username has permissions
+	if self.server.db.permissions(username, character_id) < Permission.EDIT:
+		return self.respond({"success":False, "info":MSG_NO_PERMISSION})
 	return self.respond({"success":self.server.db.remove_item(username, args[KEY_ID], args[KEY_ID2])})
 
 
 def add_item(self, args, username):
-	# TODO assert username has permissions to character
+	if self.server.db.permissions(username, character_id) < Permission.ADD:
+		return self.respond({"success":False, "info":MSG_NO_PERMISSION})
 	
 	name = args[KEY_NAME]
